feat(InputView): InputView 클래스를 프로덕션으로 이동
